+++ solution/dynamox-challenge/src/data/db.json
@@ -1,14 +1,14 @@
 {
   "users": [
     {
-      "id": "1",
+      "id": 1,
       "email": "admin@dynamox.com",
       "username": "Administrator",
       "password": "admin123",
       "role": "admin"
     },
     {
-      "id": "2",
+      "id": 2,
       "email": "user1@dynamox.com",
       "username": "Usuário 1",
       "password": "user123",
@@ -17,19 +17,19 @@
   ],
   "machines": [
     {
-      "id": "machine1",
+      "id": "ad2a1300-1888-11f0-af90-37fdd56c15c1",
       "name": "Máquina 1",
       "type": "Pump",
       "userId": 1
     },
     {
-      "id": "machine2",
+      "id": "cf4686d0-1888-11f0-af90-37fdd56c15c1",
       "name": "Máquina 2",
       "type": "Fan",
       "userId": 1
     },
     {
-      "id": "machine3",
+      "id": "ee3b7be0-1888-11f0-af90-37fdd56c15c1",
       "name": "Máquina 3",
       "type": "Pump",
       "userId": 2
@@ -37,22 +37,19 @@
   ],
   "sensors": [
     {
-      "id": "sensor1",
+      "id": "7bb4a2e0-1888-11f0-af90-37fdd56c15c1",
       "name": "Sensor 1",
-      "model": "TcAg",
-      "machineId": "machine1"
+      "model": "TcAg"
     },
     {
-      "id": "sensor2",
+      "id": "8f9c7620-1888-11f0-af90-37fdd56c15c1",
       "name": "Sensor 2",
-      "model": "TcAs",
-      "machineId": "machine2"
+      "model": "TcAs"
     },
     {
-      "id": "sensor3",
+      "id": "9cf44b40-1888-11f0-af90-37fdd56c15c1",
       "name": "Sensor 3",
-      "model": "HF+",
-      "machineId": "machine3"
+      "model": "HF+"
     }
   ]
 }
