+++ solution/dynamox-challenge/src/pages/Login.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from '../redux/reducers';
 import { useNavigate } from 'react-router-dom';
-import { loginUser } from '../redux/actions';
+import { loginUser } from '../redux/actions/loginActions';
 import { AppDispatch } from '../redux/store';
 
 export default function Login() {
