+++ solution/dynamox-challenge/src/redux/actions/machineActions.ts
@@ -1,31 +1,14 @@
-// Definição das actions
+// Actions das máquinas
 import { Dispatch } from 'redux';
-import { Machine } from '../../types';
-import { Action } from '../reducers';
+import { Machine, Action } from '../../types';
 import { ThunkAction } from 'redux-thunk';
 import { RootState } from '../store';
-import { v4 as uuidv4 } from 'uuid';
+import { v1 as uuidv1 } from 'uuid';
+
+// URL da API
 const API_UTL = 'http://localhost:3001';
 
-// Usuário
-export const loginUser = (email: string, password: string): ThunkAction<void, RootState, unknown, Action> => {
-  return async (dispatch: Dispatch) => {
-    console.log('Dispatching LOGIN_REQUEST');
-    dispatch({ type: 'LOGIN_REQUEST', payload: { email, password } });
-    
-    try {
-      const response = await fetch(`${API_UTL}/users?email=${email}&password=${password}`);
-      const data = await response.json();
-      console.log('Dispatching GET_USER with data:', data[0]);
-      dispatch({ type: 'GET_USER', payload: data[0] });
-    } catch (error) {
-      console.error('Error fetching user:', error);
-    }
-  };
-};
-
-// Máquinas
-export const fetchMachines = (
+export const getMachines = (
   userId: number | null
 ): ThunkAction<void, RootState, unknown, Action> => {
   return async (dispatch: Dispatch<Action>) => {
@@ -44,7 +27,7 @@ export const fetchMachines = (
 export const postMachine = (machine: Machine): ThunkAction<void, RootState, unknown, Action> => {
   return async (dispatch: Dispatch<Action>) => {
     try {
-      const newId = uuidv4();
+      const newId = uuidv1();
       const response = await fetch(`${API_UTL}/machines`, {
         method: 'POST',
         headers: {
@@ -106,17 +89,4 @@ export const updateMachine = (machine: Machine): ThunkAction<void, RootState, un
       console.error('Error updating machine:', error);
     }
   }
-}
-
-// Sensores
-export const fetchSensor = (id: string) => {
-  return async (dispatch: Dispatch) => {
-    try {
-      const response = await fetch(`${API_UTL}/sensors/${id}`);
-      const data = await response.json();
-      dispatch({ type: 'GET_SENSOR', payload: data });
-    } catch (error) {
-      console.error('Error fetching sensor:', error);
-    }
-  };
 }
