+++ solution/dynamox-challenge/src/redux/reducers/index.ts
@@ -1,4 +1,5 @@
 import { AuthState, Machine, Sensor } from "../../types";
+import { Action } from "../../types";
 
 export interface RootState extends AuthState {
   role: string;
@@ -16,16 +17,6 @@ const INITIAL_STATE: RootState = {
   machines: [], 
   sensors: [], 
 };
-
-export type Action =
-  | { type: "GET_USER"; payload: { id: number, username: string; role: string, machines: Machine[], sensors: Sensor[] } }
-  | { type: "GET_MACHINES"; payload: Machine[] }
-  | {type: "POST_MACHINE"; payload: Machine}
-  | { type: "DELETE_MACHINE"; payload: Machine }
-  | { type: "UPDATE_MACHINE"; payload: Machine }
-  | { type: "LOGIN_ERROR"; payload: string }
-  | { type: "LOGIN_REQUEST"; payload: { email: string; password: string } }
-  | { type: "LOGOUT" }
 
 export const rootReducer = (state = INITIAL_STATE, action: Action): RootState => {
   switch (action.type) {
