+++ solution/dynamox-challenge/src/types/index.ts
@@ -7,12 +7,11 @@ export interface AuthState {
   isLoggedIn: boolean,
 }
 
-// Máquina
 export type SensorModel = "TcAg" | "TcAs" | "HF+";
 
 export interface Sensor {
   name: string,
-  id?: string,
+  id: string,
   model: SensorModel
 }
 
@@ -21,7 +20,7 @@ export interface Machine {
   id?: string,
   type: string,
   userId: number | null,
-  sensors?: Sensor[],
+  monitoringPoints?: MonitoringPoint[],
 }
 
 // Pontos de Monitoramento
@@ -29,7 +28,17 @@ export interface MonitoringPoint {
   id?: string,
   name: string,
   machineId: string,
-  sensor: Sensor
+  sensorId: string
 }
 
-// define os tipos de dados 
+export type Action =
+  | { type: "GET_USER"; payload: { id: number, username: string; role: string, machines: Machine[], sensors: Sensor[] } }
+  | { type: "GET_MACHINES"; payload: Machine[] }
+  | {type: "POST_MACHINE"; payload: Machine}
+  | { type: "DELETE_MACHINE"; payload: Machine }
+  | { type: "UPDATE_MACHINE"; payload: Machine }
+  | { type: "LOGIN_ERROR"; payload: string }
+  | { type: "LOGIN_REQUEST"; payload: { email: string; password: string } }
+  | { type: "LOGOUT" }
+
+// define os tipos de dados e actions
